change custom wifi folder name, refactoring
